Separated frontend (#831)
* Public page and dashboard were separated into two independent builds. generating two smaller bundles.

* Public page using only MUIv5

* Teste pre-commit

* Interface de predicao funcionando

* Removed all dependencies from MaterialUI v4 and a ton of stylesheet

* enabled isort

* Fixed build

* Fixed bug in predict pipeline permission denied dates.txt

* Fixed black lint

+++ predict_occultation/src/predict_occultation/pipeline/generate_dates.py
@@ -5,33 +5,32 @@ import subprocess
 
 
 def generate_dates_file(start_date, final_date, step, filename):
-    print(f"Generate dates file: start:[{start_date}] final: [{final_date}]")
 
     app_path = os.environ.get("APP_PATH").rstrip("/")
     data_dir = os.environ.get("DIR_DATA").rstrip("/")
     output = os.path.join(data_dir, filename)
 
-    with open(output, "w") as fp:
-
-        # Py2
-        # parameters = [start_date, final_date, step]
-        # strParameters = '\n'.join(map(str, parameters))
-        # p = subprocess.Popen('geradata', stdin=subprocess.PIPE, stdout=fp)
-        # p.communicate(strParameters)
-
-        # Py3
-        parameters = [start_date, final_date, step]
-        strParameters = "\n".join(map(str, parameters))
-        p = subprocess.Popen("geradata", stdin=subprocess.PIPE, stdout=fp, shell=True)
-        p.communicate(str.encode(strParameters))
+    if not os.path.exists(output):
+        print(
+            f"Generating dates file: start:[{start_date}] final: [{final_date} step: [{step}]]"
+        )
+        with open(output, "w") as fp:
+            parameters = [start_date, final_date, step]
+            strParameters = "\n".join(map(str, parameters))
+            p = subprocess.Popen(
+                "geradata", stdin=subprocess.PIPE, stdout=fp, shell=True
+            )
+            p.communicate(str.encode(strParameters))
+    else:
+        print("Using previously created date file.")
 
     os.chmod(output, 0o664)
 
     if os.path.exists(output):
-
         # Cria um link simbolico no diretório app
-        os.symlink(output, os.path.join(app_path, filename))
-
+        output_link = os.path.join(app_path, filename)
+        os.symlink(output, output_link)
+        print(f"Created symbolic link to the dates file: [{output_link}]")
         return output
     else:
         raise (Exception("Date file not generated. [%s]" % output))
